Conflicts:
	modules/juce_opengl/native/juce_OpenGLExtensions.h
	modules/juce_opengl/native/juce_OpenGL_win32.h

+++ modules/juce_graphics/geometry/juce_Point.h
@@ -151,7 +151,8 @@ public:
     */
     FloatType getAngleToPoint (Point other) const noexcept
     {
-        return static_cast<FloatType> (std::atan2 (other.x - x, y - other.y));
+        return static_cast<FloatType> (std::atan2 (static_cast<FloatType> (other.x - x),
+                                                   static_cast<FloatType> (y - other.y)));
     }
 
     /** Returns the point that would be reached by rotating this point clockwise
@@ -169,8 +170,8 @@ public:
     */
     Point<FloatType> getPointOnCircumference (float radius, float angle) const noexcept
     {
-        return Point<FloatType> (static_cast <FloatType> (x + radius * std::sin (angle)),
-                                 static_cast <FloatType> (y - radius * std::cos (angle)));
+        return Point<FloatType> (static_cast<FloatType> (x + radius * std::sin (angle)),
+                                 static_cast<FloatType> (y - radius * std::cos (angle)));
     }
 
     /** Taking this point to be the centre of an ellipse, this returns a point on its circumference.
@@ -180,8 +181,8 @@ public:
     */
     Point<FloatType> getPointOnCircumference (float radiusX, float radiusY, float angle) const noexcept
     {
-        return Point<FloatType> (static_cast <FloatType> (x + radiusX * std::sin (angle)),
-                                 static_cast <FloatType> (y - radiusY * std::cos (angle)));
+        return Point<FloatType> (static_cast<FloatType> (x + radiusX * std::sin (angle)),
+                                 static_cast<FloatType> (y - radiusY * std::cos (angle)));
     }
 
     /** Returns the dot-product of two points (x1 * x2 + y1 * y2). */
@@ -204,13 +205,13 @@ public:
 
     //==============================================================================
     /** Casts this point to a Point<int> object. */
-    Point<int> toInt() const noexcept                             { return Point<int> (static_cast <int> (x), static_cast<int> (y)); }
+    Point<int> toInt() const noexcept                             { return Point<int> (static_cast<int> (x), static_cast<int> (y)); }
 
     /** Casts this point to a Point<float> object. */
-    Point<float> toFloat() const noexcept                         { return Point<float> (static_cast <float> (x), static_cast<float> (y)); }
+    Point<float> toFloat() const noexcept                         { return Point<float> (static_cast<float> (x), static_cast<float> (y)); }
 
     /** Casts this point to a Point<double> object. */
-    Point<double> toDouble() const noexcept                       { return Point<double> (static_cast <double> (x), static_cast<double> (y)); }
+    Point<double> toDouble() const noexcept                       { return Point<double> (static_cast<double> (x), static_cast<double> (y)); }
 
     /** Returns the point as a string in the form "x, y". */
     String toString() const                                       { return String (x) + ", " + String (y); }
